+++ RELEASE_NOTES.md
@@ -230,3 +230,4 @@ _(**NOT BACKWARDS COMPATIBLE**)_
 * Fixed some stray documentation
 * Add support for explicitly configuring RS-485 mode #117 thanks sirhcel
 * Removed checks on readInputDiscretes #114 thanks kazuyatada
+* Switched to scheduled executor service instead of timer #115 thanks mkurt